+++ CHANGELOG.md
@@ -55,6 +55,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Fix interrupted scan, when the process table is full. [#832](https://github.com/greenbone/openvas-scanner/pull/832)
 - Use fchmod to change file permission instead of on open to prevent race conditions [854](https://github.com/greenbone/openvas-scanner/pull/854)
+<<<<<<< HEAD
+=======
+- Fix plugins upload [#878](https://github.com/greenbone/openvas/pull/878)
+- Fix Error Message when NVTI chache init failed [#885](https://github.com/greenbone/openvas/pull/885)
+>>>>>>> a15e58d3 (fix error msg)
 
 [21.4.3]: https://github.com/greenbone/openvas-scanner/compare/v21.4.2...openvas-21.04